Add XML download functionality to RFQDetailsPage
Implement a new downloadXml function that allows users to download the RFQ's raw XML data as a file. This feature enhances the RFQDetailsPage by providing an easy way to access and save RFQ data. Update the UI to include a "Download XML" button, improving user experience and accessibility of RFQ information.